saya anggap beres

+++ src/ui.cpp
@@ -27,32 +27,51 @@ void UI::printChoice() // Mem-print menu yang bisa digunakan
 
 void UI::execute(int input)
 {
-    switch (input)
+    if (input == 0xffff0030)
     {
-    case 0:
         kill_command = "kill " + to_string(getpid()); // Membuat string dan mendapatkan PID program
         system(kill_command.c_str());                 // Menggunakan command linux untuk kill program berdasarkan PID nya
-        break;
-    case 1:
-        status_eksekusi = 1;
-        break;
-    case 2:
-        status_eksekusi = 2;
-        break;
-    case 3:
-        status_eksekusi = 3;
-        break;
-    case 4:
-        status_eksekusi = 4;
-        break;
-    case 5:
-        status_eksekusi = 5;
-        break;
-
-    default:
-        printChoice();
-        break;
     }
+    else if (input == 0xffff0031)
+        status_eksekusi = 1;
+    else if (input == 0xffff0032)
+        status_eksekusi = 2;
+    else if (input == 0xffff0033)
+        status_eksekusi = 3;
+    else if (input == 0xffff0034)
+        status_eksekusi = 4;
+    else if (input == 0xffff0035)
+        status_eksekusi = 5;
+    else
+        status_eksekusi = 666;
+
+    // switch (input)
+    // {
+    // case 0:
+    //     kill_command = "kill " + to_string(getpid()); // Membuat string dan mendapatkan PID program
+    //     system(kill_command.c_str());                 // Menggunakan command linux untuk kill program berdasarkan PID nya
+    //     break;
+    // case 1:
+    //     status_eksekusi = 1;
+    //     break;
+    // case 2:
+    //     status_eksekusi = 2;
+    //     break;
+    // case 3:
+    //     status_eksekusi = 3;
+    //     break;
+    // case 4:
+    //     status_eksekusi = 4;
+    //     break;
+    // case 5:
+    //     status_eksekusi = 5;
+    //     break;
+
+    // default:
+    //     // printChoice();
+    //     status_eksekusi = 666;
+    //     break;
+    // }
 }
 
 void UI::getInput()
@@ -60,6 +79,8 @@ void UI::getInput()
     static int input = -1;
 
     printf("Pilihan: ");
-    scanf("%d", &input); // Meminta inputan dari user
+    scanf("%s", &input); // Meminta inputan dari user
+    // printf("pilihan.e : %x\n", input >> 5);
+    // printf("nyobak: %x\n", input);
     execute(input);
 }
